+++ R/convertUSCensusCounties.R
@@ -71,15 +71,23 @@ convertUSCensusCounties <- function(nameOnly=FALSE) {
   SPDF$stateName <- apply(SPDF@data, 1, extractState, col='Name')
   SPDF$countyName <- SPDF$NAME
   
+  # TODO:  Figure out units for ALAND and AWATER and convert to m^2
+  
+  # Guarantee that ALAND and AWATER are numeric
+  SPDF$ALAND <- as.numeric(SPDF$ALAND)
+  SPDF$AWATER <- as.numeric(SPDF$AWATER)
+  
+  # TODO:  COUNTYNS is the polygon uniqueID but what is it?
+  
   # Subset this dataframe to include only obviously useful columns
   usefulColumns <- c('NAME','ALAND','AWATER','countryCode','countryName','stateCode','stateName',
-                     'countyName','COUNTYFP')
+                     'countyName','COUNTYFP', 'COUNTYNS')
   SPDF <- SPDF[,usefulColumns]
   names(SPDF) <- c('name', 'areaLand', 'areaWater', 'countryCode', 'countryName', 
-                   'stateCode', 'stateName', 'countyName', 'countyFIPS')
+                   'stateCode', 'stateName', 'countyName', 'countyFIPS', 'COUNTYNS')
   
   # Group polygons with the same identifier (countyName)
-  SPDF <- organizePolygons(SPDF, uniqueID='countyName', sumColumns=c('areaLand','areaWater'))
+  SPDF <- organizePolygons(SPDF, uniqueID='COUNTYNS', sumColumns=c('areaLand','areaWater'))
   
   # Assign a name and save the data
   assign(datasetName,SPDF)